+++ crates/sdk/libs/near/src/types.rs
@@ -1,6 +1,3 @@
-use serde_with::base64::Base64;
-use serde_with::serde_as;
-
 pub type BlockHeight = u64;
 pub type BlockHash = calimero_primitives::hash::Hash;
 pub type AccountId = near_account_id::AccountId;
@@ -14,45 +11,4 @@ pub type ShardId = u64;
 pub enum BlockId {
     Height(BlockHeight),
     Hash(BlockHash),
-}
-
-#[serde_as]
-#[derive(serde::Deserialize, Clone, Debug)]
-#[serde(transparent)]
-pub struct StoreValue(#[serde_as(as = "Base64")] pub Box<[u8]>);
-
-#[serde_as]
-#[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
-#[serde(transparent)]
-pub struct StoreKey(#[serde_as(as = "Base64")] pub Box<[u8]>);
-
-#[serde_as]
-#[derive(serde::Serialize, Clone, Debug)]
-#[serde(transparent)]
-pub struct FunctionArgs(#[serde_as(as = "Base64")] Box<[u8]>);
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum BlockReference {
-    BlockId(BlockId),
-    Finality(Finality),
-    SyncCheckpoint(SyncCheckpoint),
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SyncCheckpoint {
-    Genesis,
-    EarliestAvailable,
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Default, Clone, Debug)]
-pub enum Finality {
-    #[serde(rename = "optimistic")]
-    None,
-    #[serde(rename = "near-final")]
-    DoomSlug,
-    #[serde(rename = "final")]
-    #[default]
-    Final,
 }
