+++ crates/sdk/libs/near/src/views.rs
@@ -4,9 +4,7 @@ use std::sync::Arc;
 use serde_with::base64::Base64;
 use serde_with::serde_as;
 
-use crate::types::{
-    AccountId, Balance, BlockHeight, FunctionArgs, Nonce, StorageUsage, StoreKey, StoreValue,
-};
+use crate::types::{AccountId, Balance, BlockHeight, BlockId, Nonce, StorageUsage};
 
 #[derive(serde::Serialize, Debug)]
 #[serde(tag = "request_type", rename_all = "snake_case")]
@@ -108,4 +106,45 @@ pub struct AccessKeyInfoView {
 pub struct CallResult {
     pub result: Box<[u8]>,
     pub logs: Box<[Box<str>]>,
+}
+
+#[serde_as]
+#[derive(serde::Deserialize, Clone, Debug)]
+#[serde(transparent)]
+pub struct StoreValue(#[serde_as(as = "Base64")] pub Box<[u8]>);
+
+#[serde_as]
+#[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
+#[serde(transparent)]
+pub struct StoreKey(#[serde_as(as = "Base64")] pub Box<[u8]>);
+
+#[serde_as]
+#[derive(serde::Serialize, Clone, Debug)]
+#[serde(transparent)]
+pub struct FunctionArgs(#[serde_as(as = "Base64")] Box<[u8]>);
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum BlockReference {
+    BlockId(BlockId),
+    Finality(Finality),
+    SyncCheckpoint(SyncCheckpoint),
+}
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum SyncCheckpoint {
+    Genesis,
+    EarliestAvailable,
+}
+
+#[derive(serde::Serialize, serde::Deserialize, Default, Clone, Debug)]
+pub enum Finality {
+    #[serde(rename = "optimistic")]
+    None,
+    #[serde(rename = "near-final")]
+    DoomSlug,
+    #[serde(rename = "final")]
+    #[default]
+    Final,
 }
